Change count type

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: bazaluga <bazaluga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/08 16:39:36 by bazaluga          #+#    #+#             */
-/*   Updated: 2024/02/12 13:11:17 by bazaluga         ###   ########.fr       */
+/*   Updated: 2024/02/12 13:26:53 by bazaluga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,21 +52,21 @@ typedef struct s_flags
 typedef struct s_buffer
 {
 	t_type			type;
-	size_t			len;
+	int				len;
 	void			*content;
 	struct s_buffer	*next;
 }				t_buffer;
 
 /****************************** BUFFER FUNCTIONS ******************************/
 
-t_buffer	*buff_new(t_type type, size_t len, void *content);
+t_buffer	*buff_new(t_type type, int len, void *content);
 bool		buff_add_back(t_buffer **buff, t_buffer *new);
 void		*buff_clear(t_buffer **buff);
-size_t		buff_update_len(size_t to_add);
+int			buff_update_len(size_t to_add);
 
 /************************* FLAGS PARSING FUNCTIONS ****************************/
 
-void		flags_get(t_flags **flags, char *str, size_t *i, va_list args);
+void		flags_get(t_flags **flags, char *str, int *i, va_list args);
 
 /****************************** UTILS FUNCTIONS *******************************/
